+++ packages/sdk/src/v2openapi/index.ts
@@ -48,7 +48,7 @@ import {
 import { EnvironmentVariable, InputVariable, mapInputVariable } from '../configure/variables';
 import { HeadersBuilder, mapHeaders } from '../definition/headers-builder';
 import { Logger } from '../logger';
-import _ from 'lodash';
+import { camelCase } from 'lodash';
 import transformSchema from '../transformations/transformSchema';
 
 export const openApiSpecificationToRESTApiObject = async (
@@ -1209,8 +1209,8 @@ class RESTApiBuilder {
 
 	private cleanupTypeName = (typeName: string, parentTypeName: string): string => {
 		// remove all non-alphanumeric characters and all leading numbers
-		typeName = _.camelCase(typeName.replace(/[^_a-zA-Z0-9]/g, '_').replace(/^[0-9]+/, '_'));
-		parentTypeName = _.camelCase(parentTypeName.replace(/[^_a-zA-Z0-9]/g, '_').replace(/^[0-9]+/, '_'));
+		typeName = camelCase(typeName.replace(/[^_a-zA-Z0-9]/g, '_').replace(/^[0-9]+/, '_'));
+		parentTypeName = camelCase(parentTypeName.replace(/[^_a-zA-Z0-9]/g, '_').replace(/^[0-9]+/, '_'));
 		// and make the first character uppercase
 		typeName = typeName[0].toUpperCase() + typeName.substring(1);
 		parentTypeName = parentTypeName[0].toUpperCase() + parentTypeName.substring(1);