+++ packages/solid-query/tests/index.test-d.ts
@@ -1,5 +1,5 @@
 import { createHooks } from '../src';
-import { Client, ClientResponseError, OperationsDefinition, User } from '@wundergraph/sdk/client';
+import { Client, ResponseError, OperationsDefinition, User } from '@wundergraph/sdk/client';
 import { expectType } from 'tsd';
 import { CreateQueryResult } from '@tanstack/solid-query';
 
@@ -9,7 +9,7 @@ interface Operations extends OperationsDefinition {
 			input: {
 				city: string;
 			};
-			data: any;
+			response: { data?: { temperature: number }; error?: ResponseError };
 			requiresAuthentication: boolean;
 		};
 	};
@@ -18,7 +18,7 @@ interface Operations extends OperationsDefinition {
 			input: {
 				forCity: string;
 			};
-			data: any;
+			response: { data?: { temperature: number }; error?: ResponseError };
 			requiresAuthentication: boolean;
 		};
 	};
@@ -27,7 +27,7 @@ interface Operations extends OperationsDefinition {
 			input: {
 				name: string;
 			};
-			data: any;
+			response: { data?: { name: string }; error?: ResponseError };
 			requiresAuthentication: boolean;
 		};
 	};
@@ -49,8 +49,8 @@ const { data: queryData, error: queryError } = createQuery({
 	},
 });
 
-expectType<Operations['queries']['Weather']['data']>(queryData);
-expectType<ClientResponseError | null>(queryError);
+expectType<Operations['queries']['Weather']['response']['data']>(queryData);
+expectType<ResponseError | undefined | null>(queryError);
 
 const { data: subData, error: subError } = createSubscription({
 	enabled: true,
@@ -61,8 +61,8 @@ const { data: subData, error: subError } = createSubscription({
 	},
 });
 
-expectType<Operations['subscriptions']['Weather']['data']>(subData);
-expectType<ClientResponseError | null>(subError);
+expectType<Operations['subscriptions']['Weather']['response']['data']>(subData);
+expectType<ResponseError | undefined | null>(subError);
 
 const {
 	data: mutData,
@@ -73,8 +73,8 @@ const {
 	operationName: 'CreateUser',
 });
 
-expectType<Operations['mutations']['CreateUser']['data']>(mutData);
-expectType<ClientResponseError | null>(mutError);
+expectType<Operations['mutations']['CreateUser']['response']['data']>(mutData);
+expectType<ResponseError | undefined | null>(mutError);
 
 expectType<void>(
 	mutate({
@@ -82,14 +82,14 @@ expectType<void>(
 	})
 );
 
-expectType<Promise<any>>(
+expectType<Promise<Operations['mutations']['CreateUser']['response']['data']>>(
 	mutateAsync({
 		name: 'John Doe',
 	})
 );
 
-expectType<CreateQueryResult<User<string>, ClientResponseError>>(useUser());
-expectType<CreateQueryResult<User<string>, ClientResponseError>>(
+expectType<CreateQueryResult<User<string>, ResponseError>>(useUser());
+expectType<CreateQueryResult<User<string>, ResponseError>>(
 	useUser({
 		revalidate: true,
 		abortSignal: new AbortController().signal,