+++ testapps/default/test/index.test.ts
@@ -4,22 +4,27 @@ import { createTestServer } from '../.wundergraph/generated/testing';
 
 const wg = createTestServer();
 
-// TODO: Remove this once the secret refactor is done
-process.env['WG_SECURE_COOKIE_HASH_KEY'] = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
-process.env['WG_SECURE_COOKIE_BLOCK_KEY'] = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
-process.env['WG_CSRF_TOKEN_SECRET'] = 'aaaaaaaaaaa';
-
 beforeAll(() => wg.start());
 afterAll(() => wg.stop());
 
 describe('functions', () => {
 	test('internal operation call from  function', async () => {
 		const client = wg.client();
-		const result = await client.query({
-			operationName: 'nested/InternalWrapper',
-		});
-		expect(result.error).toBeUndefined();
-		expect(result.data?.data?.chinook_findFirstAlbum?.AlbumId).toBe(1);
+		const promises = [];
+		// Call this 50 times to exercise some code paths that share cached
+		// buffers. If we have a race condition in there, the go race detector
+		// (which we use in CI) will likely catch it.
+		for (let ii = 0; ii < 50; ii++) {
+			const op = client.query({
+				operationName: 'nested/InternalWrapper',
+			});
+			promises.push(op);
+		}
+		const results = await Promise.all(promises);
+		for (const result of results) {
+			expect(result.error).toBeUndefined();
+			expect(result.data?.data?.chinook_findFirstAlbum?.AlbumId).toBe(1);
+		}
 	});
 });
 
@@ -141,5 +146,34 @@ describe('sql', () => {
 		expect(result.data?.row[0].id).toEqual(2);
 		expect(result.data?.row[0].name).toEqual('Jannik');
 		expect(result.data?.row[0].email).toEqual('jannik@wundergraph.com');
+	});
+});
+
+describe('@transform directive', () => {
+	test('@transform in public operation', async () => {
+		const result = await wg.client().query({
+			operationName: 'CountryWeather',
+			input: {
+				countryCode: 'ES',
+			},
+		});
+		expect(result.error).toBeUndefined();
+		const country = result?.data?.country?.[0];
+		expect(country).toBeDefined();
+		expect(country?.capital).toBe('Madrid');
+		expect(country?.weather.temperature?.max).toBeDefined();
+	});
+	test('@transform in internal operation', async () => {
+		const result = await wg.client().query({
+			operationName: 'CountryWeatherViaInternal',
+			input: {
+				countryCode: 'ES',
+			},
+		});
+		expect(result.error).toBeUndefined();
+		const country = result?.data?.data?.country?.[0];
+		expect(country).toBeDefined();
+		expect(country?.capital).toBe('Madrid');
+		expect(country?.weather.temperature?.max).toBeDefined();
 	});
 });