+++ packages/golang-client/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.5.0...@wundergraph/golang-client@0.5.1) (2023-03-29)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.5.0](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.4.9...@wundergraph/golang-client@0.5.0) (2023-03-28)
+
+### Features
+
+* custom error support for typescript operations ([#759](https://github.com/wundergraph/wundergraph/issues/759)) ([7cfa815](https://github.com/wundergraph/wundergraph/commit/7cfa815a44b0b2a1184c2c50d0c4ea7cbb8066a7)) (@StarpTech)
+
+## [0.4.9](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.4.8...@wundergraph/golang-client@0.4.9) (2023-03-27)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
 ## [0.4.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.4.7...@wundergraph/golang-client@0.4.8) (2023-03-23)
 
 **Note:** Version bump only for package @wundergraph/golang-client